requirements.txt
instagrapi==1.17.4
python-dotenv==1.0.1
python-telegram-bot==20.7
Pillow>=8.1.1
urllib3<2.0